+++ src/components/stats/types.ts
@@ -1,6 +1,6 @@
 import { type Accessor } from "solid-js";
 import type { SpotifyApi } from "@spotify/web-api-ts-sdk";
-import type { ActualPlaylistedTrack } from "../../SpotifyHelper";
+import type { ActualPlaylistedTrack } from "../../types/spotify";
 
 export type StatsComponentProps = {
   spotify: Accessor<SpotifyApi | null>;
